requirements.txt
gradio==4.22.0
gradio-imageslider==0.0.16
pygltflib==1.16.1
trimesh==4.0.5
imageio
imageio-ffmpeg
Pillow

accelerate>=0.22.0
diffusers==0.27.2
matplotlib==3.8.2
scipy==1.11.4
torch==2.0.1
transformers>=4.32.1
xformers>=0.0.21
